+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.3] - 2022/02/17
+### Changed
+- Changed type of options in toXml() method to XmlFormattingOptions interface.
+### Added
+- Added includeProcessingInstructions option to XmlFormattingOptions.
+
 ## [0.1.2] - 2022/01/29
 ### Added
 - Added `findChild()` method on XmlNode.